Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-float a;
+double a;
 printf("enter a double precesion floating point number:");
 scanf("%1f",&a);
-printf("You entered: %.61f\n", a);
+printf("You entered: %.41f\n", a);
 return 0;
 }
